# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "API documentation for Task Manager application",
        "contact": {
            "name": "Valiavo",
            "email": "valiavoandriantsoa30@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "APIs Server"
        }
    ],
    "paths": {
        "/v1/tasks": {
            "get": {
                "tags": [
                    "Tasks V1"
                ],
                "summary": "Liste de toutes les tâches",
                "operationId": "cef450ac078bbbd790a9b74cad86791a",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks V1"
                ],
                "summary": "Créer une nouvelle tâche",
                "operationId": "970a65037b4a2dfd6481ba305e8f606b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Ma tâche"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description de la tâche"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tâche créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/v1/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks V1"
                ],
                "summary": "Obtenir une tâche spécifique",
                "operationId": "2f173fadd6c381d3758d7a0578e660e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tâche non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks V1"
                ],
                "summary": "Mettre à jour une tâche",
                "operationId": "21d3a613f3fb1bcdbcd10b0a7f9b0438",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Tâche mise à jour"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description mise à jour"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tâche mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tâche non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks V1"
                ],
                "summary": "Supprimer une tâche",
                "operationId": "716e78a2aaa5ce9eeaa3e2166c86375c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tâche supprimée avec succès"
                    },
                    "404": {
                        "description": "Tâche non trouvée"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "required": [
                    "titre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "titre": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "completed": {
                        "type": "boolean",
                        "default": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TaskRequest": {
                "required": [
                    "titre"
                ],
                "properties": {
                    "titre": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "completed": {
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks V1",
            "description": "Tasks V1"
        }
    ]
}